pyproject.toml
[project]
name = "date-stub-generator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "convertdate>=2.4.0",
    "ipykernel>=7.0.1",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "pydantic>=2.12.2",
    "python-dotenv>=1.1.1",
    "streamlit>=1.50.0",
]
